app/requirements.txt
django==1.3
south
django-tagging
django-piston